+++ include/ignition/gazebo/components/Transparency.hh
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 Open Source Robotics Foundation
+ * Copyright (C) 2020 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,11 @@
  * limitations under the License.
  *
  */
-#ifndef IGNITION_GAZEBO_COMPONENTS_JOINTPOSITION_HH_
-#define IGNITION_GAZEBO_COMPONENTS_JOINTPOSITION_HH_
-
-#include <vector>
+#ifndef IGNITION_GAZEBO_COMPONENTS_TRANSPARENCY_HH_
+#define IGNITION_GAZEBO_COMPONENTS_TRANSPARENCY_HH_
 
 #include <ignition/gazebo/components/Factory.hh>
 #include <ignition/gazebo/components/Component.hh>
-#include <ignition/gazebo/components/Serialization.hh>
 #include <ignition/gazebo/config.hh>
 
 namespace ignition
@@ -32,13 +29,12 @@ namespace gazebo
 inline namespace IGNITION_GAZEBO_VERSION_NAMESPACE {
 namespace components
 {
-  /// \brief Joint positions in SI units (rad for revolute, m for prismatic).
-  /// The component wraps a std::vector of size equal to the degrees of freedom
-  /// of the joint.
-  using JointPosition = Component<std::vector<double>, class JointPositionTag,
-                                  serializers::VectorDoubleSerializer>;
-  IGN_GAZEBO_REGISTER_COMPONENT(
-      "ign_gazebo_components.JointPosition", JointPosition)
+  /// \brief A component used to indicate an entity's transparency value
+  /// e.g. visual entities. Value is in the range from 0 (opaque) to
+  /// 1 (transparent).
+  using Transparency = Component<float, class TransparencyTag>;
+  IGN_GAZEBO_REGISTER_COMPONENT("ign_gazebo_components.Transparency",
+      Transparency)
 }
 }
 }